String comparisons are slow.  This new caching greatly improved map loop speed.
In my environment, this dropped us from around 4 minutes for all() 140K rows
in session queries to closer to 3 minutes.  The same query took about 12
seconds when constructed directly from the lower level metadata API.

I saw room for more improvements like this with regard to functions performed
in the critical loop (the map { callback } @$results in the iterator).  This
is a critical loop.  On the other hand, if someone is fetching 140K rows as
mapped objects, they might be approaching the task incorrectly.

Another thought is to extend the API a little:

  $query->execute->all_arrayref  # make all() thin by calling this and then map
  $query->execute->all_hashref   # derive column names, make hashes, no bless

Experiments revealed that this kind of to-the-metal query only takes 3 seconds.