Added more information about the original table per column.
Although I may alias my tables, I still want to know the original table of
each of the columns.  This is useful for relation-based typing, wherein I need
to get the header ($query->builder->column) to further interpret the results.

Side note: It is inconsistent that a different storage strategy is used for
column aliases vs table aliases.