Select queries now support add_inner_join() and add_left_join().
The two new functions serve a few purposes.

First, they follow the principle of least surprise.  The join() and add_join()
functions default to "LEFT OUTER" joins, as opposed to SQL default of "INNER".
Since their names are explicit, a user knows exactly what will be added to the
join clauses.

Second, since the join type is in the function name, the parameters may be
less awkward.  Compare:

  $query->add_join(
      [$his_table => [$id == 1], 'INNER'],
      [$her_table => [$id == 2], 'INNER'],
  )

to:

  $query->add_inner_join(
      $his_table => [$id == 1],
      $her_table => [$id == 2],
  )

Third, there are really only two semantically distinct kinds of joins in SQL:
inner and left.  There is also "right", but it is rarely used and a simple
transform of the parameters.  Thus, the two functions serve the purpose of
narrowing down the options.  A user may fall back to using "add_join" instead
if other joins are important.  Also, "left" always implies "outer", so there
is no need to include the keyword.

Fourth, there is a distinct lack of inner_join() and left_join() functions.  I
noticed when people play with the query-building API, they expect fresh
instances of the method calls to ADD to the clause slots, not replace them.
So, while respecting the original API's behavior, I have intentionally
neglected the slot-replacement, especially since adds work on empty slots.  My
general recommendation to my clients will be to always use "add_", which I
believe should be a default behavior with "replace_" being the less-used but
still available behavior.