Removed Module::Find dependency.
Using Module::Find makes the system run slower with additional I/O, and it is
simply magic.  Considering that OM::Metadata::Sugar dynamically creates
functions based on the types, we are running the risk of a type with the same
name as a Perl built-in.  This entire dynamic method-building plus
module-finding tends to be a design smell.  This change is the first step
towards a safer system.

I initially ran into problems with this because I had a global version of OM
installed in addition to linking to my local repo lib, so Sugar actually found
two classes for each type and spit out warnings about function redefinition.
My first inclination was to use List::MoreUtils::uniq, but that fixed the
symptom, not the problem.